update to latest arcdps
This comes with several changes:

First, the revision header field is now properly parsed and saved,
instead of just being hardcoded to zero. This is the first step in
allowing newer log files to be parsed. To accommodate this, the Header
struct has been extended with the "revision: u8" field.

To be able to parse both formats, the CbtEvent struct has been changed.
It is now the unification of both the new struct and the old struct, as
the changes are pretty minor and mostly concern the parsing itself. The
data types have been adjusted, and two fields have been added. Moving
fields around does not concern CbtEvent at all. If the struct diverges
more from this in the future, some splitting might be introduced, but
for now, the change is pretty transparent to other code.

During this process, the structs have lost their [repr(C]) property. It
was never used, as the structs were not directly involved in C FFI. It
has been a relic of the past from earlier iterations.

Finally, the parsing function has been changed from parse_event to
parse_event_rev0, with the addition of a parse_event_rev1. parse_events
now takes the event parsing function as an additional parameter, and
parse_file correctly choses the implementation based on the revision
number.

+++ src/event.rs
@@ -56,7 +56,7 @@ pub enum EventKind {
     /// A skill has been used.
     SkillUse {
         source_agent_addr: u64,
-        skill_id: u16,
+        skill_id: u32,
         activation: Activation,
     },
 
@@ -64,7 +64,7 @@ pub enum EventKind {
     ConditionTick {
         source_agent_addr: u64,
         destination_agent_addr: u64,
-        condition_id: u16,
+        condition_id: u32,
         damage: i32,
     },
 
@@ -72,14 +72,14 @@ pub enum EventKind {
     InvulnTick {
         source_agent_addr: u64,
         destination_agent_addr: u64,
-        condition_id: u16,
+        condition_id: u32,
     },
 
     /// Physical damage.
     Physical {
         source_agent_addr: u64,
         destination_agent_addr: u64,
-        skill_id: u16,
+        skill_id: u32,
         damage: i32,
         result: raw::CbtResult,
     },
@@ -88,16 +88,16 @@ pub enum EventKind {
     BuffApplication {
         source_agent_addr: u64,
         destination_agent_addr: u64,
-        buff_id: u16,
+        buff_id: u32,
         duration: i32,
-        overstack: u16,
+        overstack: u32,
     },
 
     /// Buff removed.
     BuffRemove {
         source_agent_addr: u64,
         destination_agent_addr: u64,
-        buff_id: u16,
+        buff_id: u32,
         total_duration: i32,
         longest_stack: i32,
         removal: raw::CbtBuffRemove,
